Implementa análisis de reviews y método para actualizar la puntuación de los produdctos

+++ models/product.go
@@ -3,5 +3,5 @@ package models
 type Product struct {
 	ID             int64
 	Name           string
-	Average_Rating float64
+	Average_Rating float32
 }
